requirements.txt
appdirs==1.4.4
asgiref
certifi==2021.5.30
distlib==0.3.2
Django
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
mysqlclient
psycopg2
pytz
sqlparse
virtualenv==20.4.7
